Replace (most of) memset() with ranges::fill()
For several reasons:

In which order do the memset parameters go?
   memset(p, 10, 20);
Does this write 10 bytes with value 20, or 20 bytes with value 10? (It's
the latter.) Compare this with:
   ranges::fill(std::span{p, 20}, 10);
This might be longer to write, but it's unambiguous.

In some cases ranges::fill() is also shorter to write. Compare:
    memset(buf, 0, sizeof(buf));
    ranges::fill(buf, 0);

Memset takes the size as a number of bytes, not number of elements.
Compare:
    memset(buffer, 0, count * sizeof(float));
    ranges::fill(std::span{buffer, count}, 0.0f);
It's easy to forget to multiply with 'sizeof(T)' (for non-byte buffers).

Memset can only fill a repeating byte pattern, not repeating elements.
    ranges::fill(myFloatBuffer, 123.45f);
The float value '123.45' is probably not represented by a single
repeating byte.

Pointer vs array mistakes:
    memset(buf, 0, sizeof(buf));
This does what you want when 'buf' is an array (it uses
array-to-pointer-decay). But if 'buf' is (already) a pointer, the above
statement also compiles fine, but does probably not what you want at
runtime. The corresponding statement:
    ranges::fill(buf, 0)
will instead trigger a compile error if 'buf' is a pointer (instead of
an array).

I checked various cases, and gcc/clang generate identical x86 code for
corresponding memset() and ranges::fill() statements.