Fix the combination of (Key)Joystick and NinjaTap
Commit bb839473 fixed our implementation of (Key)Joystick to always
return 0x3f (no direction or triggers pressed) when pin8 is high. This
is correct.

On the 2 MSX joystick ports pin 8 can be controlled via software. Though
on the 4 NinjaTap joystick ports pin 8 is connected to GND, so always
low.

In openMSX in newly created joystick ports the initial status of pin 8
is high. For the MSX joystick ports this status is quickly overwritten
by the actual value (controlled via the PSG). But for the NinjaTap ports
there is no code that changes the value of the output pins, so pin 8
remains high. (Before patch bb839473 this didn't matter.)

Fixed by initializing the NinjaTap output pins to low.