Revert 0a0ad35e45: couple input/output joystick ports pin 6/7
See comments in the code for more details.

Unfortunately commit 0a0ad35e45 was made without any reference to a test
case. My *guess* is that is true for machines with a discrete AY8910
chip, but not on machines where the PSG is integrated in the MSX-engine
chip.

Reverted it for now because it does e.g. break the NinjaTap test
programs. And (unfortunately) it's unclear which tests the patch was
meant to fix.