NinjaTap: fix behavior of pin7
Thanks to Danjovic's analysis! See post from 08-10-2022, 06:58
    https://www.msx.org/forum/msx-talk/hardware/ninja-tap?page=7

That post also talks about a 'fix' for MSXPSG. Though that's more of a
hack specific for NinjaTap rather than a general fix. I think a proper
fix would be to implement real tri-state logic (with 0, 1, and Z (high
impedance) values). Let's first see how much work it is to implement
that before adding this hack.