Tweak Ram interface (and CheckedRam, SRAM, TrackedRam, Rom)
The Ram class already gives access to the full memory content via the
operator[] method. This patch extends that interface with begin()/end()
methods to make it possible to operate on this memory range via STL
algorithms.

Classes with a pair of begin()/end() methods implement the c++20 "range"
concept (that is: they can be passed to functions that require a range
as input). If they also implement a size() method then they satify the
"sized_range" concept. The Ram class already implemented a 'getSize()'
method. We renamed that to 'size()'. And to keep the code consistent, we
did the same rename in the CheckedRam, TrackedRAm, SRAM and Rom classes.

The old 'getSize()' method returned an 'unsigned' (32 bit). The renamed
'size()' method returns a 'size_t' (32 or 64 bit depending on the CPU).
Not that we'll likely need more than 32 bit for the (emulated) Ram size.
But (because of alignment/padding) storing a size_t does not actually
increase the size of the Ram object. And using 'size_t' instead of
'unsigned' in array-index operations does actually generate (a tiny bit)
better x86_64 assembly code (we don't have to explicitly truncate
calculated values to 32 bit).