added ranges::copy(sized_range, sized_range)
Our ranges::copy() (modeled after c++20 std::ranges::copy()) takes a
"range" as the source parameter and an output-iterator as the
destination parameter. Often using this algorithm is more convenient
when both source and destination are ranges. Compare:
    ranges::copy(from, to.data()); // range + output-iterator
    ranges::copy(from, to);        // range + range
So this patch adds that 2nd overload.

An extra advantage is that the 2nd form allows to assert that the
destination must be large enough to hold all elements of the source. In
other words the size of the destination range must be at least as large
as the size of the source range. Though this can only be checked for
"sized ranges" (that is: the size can be calculated without consuming
the range, a counter-example would be reading from stdin). Note:
currently it's actually only implemented for sized ranges, but this
could be extended in the future.

Adding this new overload is not 100% backwards compatible with existing
users of ranges::copy(). For details see the section "bw-compat" in the
new unittest. I've fixed these problems and marked them with "TODO
simplify", this will be addressed in later patches.

This patch only introduces the new overload but doesn't use it yet.
That's for later patches.