Added subspan() utility
This patch only adds the utility (+ unittest). Later patches will use
it.

--

Some of the reasons for why c++ is a memory UNsafe language are:
* Buffer overflows (e.g. index out-of-range).
* Pointer arithmetic (several problems, but in this context: the newly
  calculated pointers can easily point out-of-bound).
The following 100+(!) patches (mostly many small ones) are an
experiment: we change the coding style to make the above problems less
likely. And in fact we caught a real bug with it (see later patch about
TigerTreeHash).

The main idea is the following:
* Don't use pointers to point to more than one object (e.g. the start of
  an array). Instead use std::span for that.

Example 1:
    void process(Foo* buffer, size_t count);
Probably 'buffer' points to the start of an array that contains 'count'
number of elements. We can make this more explicit by changing the
function to:
    void process(std::span<Foo> buffer);
And this has the advantage that the compiler (or the library)
'understands' what's going on. E.g. in debug mode we can now do
bound-checks.

Example 2:
    void renderScreen5Line(const byte* vramLine, Pixel* buffer);
When rendering a screen-5-line we need 128 bytes of VRAM and that will
produce 256 host pixels. So probably 'vramLine' points to the start of a
(read-only) 128-byte region. And then 'buffer' must point to the start
of a 256 host pixel buffer. This can be communicated like:
    void renderScreen5Line(std::span<const byte, 128> vramLine,
                           std::span<Pixel, 256> buffer);
This much better expresses the intent. And in debug mode the compiler
can now do bound-checks.

Example 3:
    void scale1on2(const Pixel* in, Pixel* out, size_t n);
This function is part of LineScaler.hh. It doubles every pixel from 'in'
into 'out'. Though from this signature it's not clear whether 'n'
denotes the size of the input or the output buffer. An alternative could
be:
    void scale1on2(std::span<const Pixel> in, std::span<Pixel> out);
with this assert in the implementation of the function:
    assert(2 * in.size() == out.size());
And once again, this allows to do bounds-checks (in debug mode).

Some notes on efficiency:
* Bounds checks are optional (e.g. only in debug mode). Though in some
  cases they can already happen at compile-time (so no runtime cost, but
  extra safety for free).
* A std::span comes in two flavors: dynamically or statically sized:
  - Example 1 use dynamically sized spans. Internally these store both a
    pointer and a size. So example 1 possibly generates identical code
    for both versions of the function. In other words: using span has
    zero overhead.
  - Example 2 uses fixed-size spans. The size is part of the type, it's
    not stored at-run-time. So also example 2 is zero overhead.
* Example 3 does introduce a little overhead:
  - Before only one size parameter was passed and this conveyed the size
    of both buffers.
  - After the rewrite we pass two size parameters (as part of the span).
    This is redundant (but it does allow to verify the sizes in debug
    mode). To be clear we reduced the number of parameters from 3 to 2,
    but these 2 new parameters each contain to pieces of information. So
    we effectively pass 4 'words of information (pointers of size)'.
    Depending on the platform/compiler, passing this extra information
    is done via registers, so very cheap. In other words: using
    std::span here is not _zero_ overhead, but the extra cost is
    negligible.

In summary:
Using std::span results in more expressive code. The compiler/library
can do better checks (compile-time, or optionally at run-time). And
there's (practically) no run-time overhead (in non-debug mode).

--

To achieve the above goal we largely rely on the following 3 clang-tidy
checks:
* cppcoreguidelines-pro-bounds-array-to-pointer-decay
* cppcoreguidelines-pro-bounds-pointer-arithmetic
* cppcoreguidelines-avoid-c-arrays
I mean fixing all warnings generated by these 3 checks already goes a
long way in replacing pointers+size with std::span.