BUGFIX: tiger-tree-hash (tth) calculation was broken
Commit 91097b3bb4ae9fa833e31e6340231963441964f9 tried to speedup the
calculation by increasing the block size from 1kB to 64kB. Though it had
a bug: what it actually did was hash 1kB out of every 64kB. Obviously
that can be much faster, but it's not what we want :(

When I fix the bug, it's not that much faster than the original.
Therefor I reverted to a block size of 1kB. Also because that's a more
"standard" way to implement "tth" (e.g. the "rhash" tool (only) supports
this block size).

Unfortunately the unittest that's supposed to verify the correctness of
the tth-calculation was only added _after_ the bug was introduced. And
the unittest trusted the values at that point. Today I _did_ verify the
values by manually constructing a file with the tested content and then
calculating the hash-value via the "rhash" command line tool.
    https://github.com/rhash/RHash

This does mean that calculating the initial hash for a (large) hard-disk
when starting openMSX does become quite a bit slower again. But I'm
afraid there's not much we can do about that :(

I discovered this bug by accident while working on std::span/std::array
refactorings. It's a nice demonstration of how using higher level
abstractions can result in more correct code (that is: incorrect code
often doesn't nicely fit the higher-level pattern and thus stands out).