Return a std::span from our VLA macro
A line like:
    VLA(int, myBuffer, num);
Used to expand to:
    int myBuffer[num];
(or something more complex (using alloca) on visual studio)

Now it expand to:
    int myBuffer_storage[num];
    std::span myBuffer{myBuffer_storage, num);

So the effect is that 'myBuffer' changes from a C-array (with variable
length) to a std::span. This makes using it often more convenient and/or
safer.

Because the extra 'std::span' object is a local variable on the stack,
the compiler is generally very good in optimizing it away.