fix crash in SpriteConverter in debug build (STL debug checks)
In the SpriteConverter constructor we temporarily initialized 'palette'
to the (invalid) value 'std::span{nullptr, 16}'. Because it would anyway
soon afterwards be overwritten with the correct value. (Before the
ptr->span refactoring, 'palette' was a pointer and it was left
uninitialized).

Though DEBUG-STL has stricter checks on the initialization of std::span
objects. They can only be initialized with a valid range (and {nullptr,
16} is not a valid range).

Fixed by already passing the correct palette to the SpriteConverter
constructor.