Fix harmless(?) uninitialized memory file write in AviWriter
Chunks in an AVI file must have a an even size. If we want to write an
odd number of bytes, this chunk must be padded.

Before we simply rounded-up the given size and wrote that many bytes
from the given buffer. Though this reads one byte past the given input
buffer. And (in some situations) valgrind then complains that we write
an uninitialized byte to the file.

Now we only write the given buffer, followed by a write of an extra
zero-byte.