bugfix after dfffb03f4c, Introduce 'struct StereoFloat'
This patch corrects 3 mistakes:

1) copy/paste error in MSXMixer::updateStream()
We want to conditionally (depending on whether SSE2 instructions are
available) align 'mixBuffer'. This was done via #ifdef/#endif right
before the definition of that variable. Later that variable was moved to
a different location (closer to where it's actually needed), but the
ifdef-block was not moved along. Fixed by replacing the #ifdef block
with the 'ALIGNAS_SSE' macro, this allows to keep the annotation and the
variable on the same line.
This is a mistake and it should be fixed, though likely in practice this
didn't cause problems.

2) One of the goals of patch dfffb03f4c was to simplify the code by
removing "multiply-by-2" or "divide-by-2" expression to deal with the
difference between mono or stereo, or with counts in number of
audio-framents (left + right count as 1) or in number of samples (left +
right count as 2). In the buffer-allocation code in SDLSoundDriver we
overlooked one such simplificaton. The effect was that the buffer was
now twice as large as intended (but apart from this it would just work).

3) This is an actual crash, but difficult to trigger. It requires
creating a recording in combination with an (un)lucky interaction
between the openMSX main-thread and the (SDL-)audio-thread:

When we are recording audio, and when the openMSX main-thread (= the
audio-producer) runs too fast compared to the audio-thread (= the
audio-consumer) then we will drop the excess samples.

Patch dfffb03f4c changed the interface of uploadBuffer()
  from: uploadBuffer(const float* buffer, unsigned len);
  to:   uploadBuffer(std::span<const StereoFloat> buffer);
So we don't have a 'len' variable anymore, that's now 'buffer.size()'.
Though because 'len' was used in several places I re-introduced it as a
local variable (len = buffer.size()), so far so good. The problem
occured in the "drop excess samples" behavior. There we reduce the value
of this 'len' variable. In the old code that's fine. But in the new code
this breaks the invariant that 'len == buffer.size()' (and this was
needed for some of the later buffer-copy-code).

Fixed by removing the 'len' variable and instead use 'buffer.size()'
everywhere. This is a bit longer to write, but there's no redundant
state anymore (that can get out-of-sync). And the compilers optimizer
will most likely eliminate all size() calls.