Introduce narrow() and narrow_cast()
See comments in src/utils/narrow.hh for more details. In short:
* narrow_cast(): a searchable way to do narrowing casts of values.
* narrow(): a checked version of narrow_cast(), asserts that the
            numerical value remains unchanged.

Last week I pushed a patch-series that addresses (a small part of) c++
memory safety. Another aspect of safety are integer conversions. For
example C/C++ will silently convert an 'int' to a 'char', or an 'int' to
an 'unsigned', and this might change the numerical value. Sometimes this
change in value is intended, sometimes it's a bug.

This patch adds two utilities:
 * narrow_cast: for when the change is intended
 * narrow: for when the value should not change
And the latter can be checked (only in an assert-build).

This patch adds the infrastructure for safe narrowing conversions, but
only uses it in two minor places. Places pointed out by "clang-tidy
modernize-use-auto". For example:
    uint8_t sector = uint8_t(log + 1);
The clang-tidy check was actually about the repeated use of the type
'uint8_t' (the first can be replaced with 'auto'). But it's also good to
make the intended narrowing-conversion explicit: in this case it should
not truncate the numerical value. So the line becomes:
    auto sector = narrow<uint8_t>(log + 1);

There are _LOTS_ more places where we can use these two new utilities.
We could search for them via:
 * gcc warning -Wconversion
 * visual studio warning C4267: ... conversion .. possible loss of data
 * clang-tidy bugprone-narrowing-conversions
Maybe I'll make some (incremental?) effort in the future. But for now
it's good to add the infrastructure so that we can already use this in
new code.