[1/8] Fix some clang-tidy bugprone-implicit-widening-of-multiplication-result
https://clang.llvm.org/extra/clang-tidy/checks/bugprone/implicit-widening-of-multiplication-result.html

Consider the following code fragment on a 64-bit architecture:
    char* buffer = ...;
    uint32 pitch = ...;
    for (uint32_t i : xrange(num)) {
        buffer[i * pitch] = ...;
    }

The clang-tidy check will complain about the indexing operation in this
example. It searches for multiplications that produce a narrow result
(32 bit in this example) which is then extended to a wider type (64 bit
in this example).

In this example both 'i' and 'pitch' are 32 bit. Multiplying them still
produces a 32 bit result (required by C/C++), so this operation might
overflow. This product is then extended to 64 bit before it can be added
to the 64 bit 'buffer' pointer. If we would instead first extend 'i' and
'pitch' to 64-bit and only then do the multiplication, there cannot be
an overflow.

(For quite some time already) on many 64-bit CPUs, doing a 64-bit
multiplication is just as fast as doing a 32-bit multiplication. Also
many CPUs have widening-multiplication instructions that take 2
32-bit inputs and produce a 64-bit result. C/C++ does not directly
support such widening-operations. Instead you first cast the inputs to
the wider type and multiply those. Compilers can recognize this pattern
and map it to a single CPU instruction.

So doing the widening before the multiplication has no negative effect
on performance and it might fix overflow problems. Though in practice
these overflow bugs might be rare, I don't believe there were any in
openMSX.

We can achieve this by writing the above code like:
    char* buffer = ...;
    size_t pitch = ...; // <-- changed 'uint32_t' to 'size_t'
    for (uint32_t i : xrange(num)) {
        buffer[i * pitch] = ...; // this now does a 64 bit multiplication
    }

The compiler is not allowed to do this transformation by itself because
it changes the semantics of the code: overflow on unsigned types is well
defined in C/C++ (it does modulo-wrapping). Therefor the compiler must
preserve this behavior (it's different for *signed* types, but that's
another topic).

--

In the above example the compiler can perform a "strength reduction"
optimization. That is: reduce repeated multiplications in the loop to
additions. The (original) example then becomes something like:
    char* buffer = ...;
    uint32 pitch = ...;
    uint32_t index = 0;
    for (uint32_t i : xrange(num)) {
        buffer[index] = ...;
        index += pitch;
    }

Note that this transformation did NOT change the semantics. The
potential overflow is still present ('index' is still a 32-bit value,
and repeated additions can still overflow in exactly the same way).

Also note that the subexpression 'buffer[index]' must still first expand
'index' to 64-bit before it can be added to the 'buffer' pointer (most
CPUs don't directly support mixed 32+64 bit addition).

If we instead perform strength-reduction on the 2nd example, we can get
something like:
    char* buffer = ...;
    size_t pitch = ...;
    for (uint32_t i : xrange(num)) {
        *buffer = ...;
        buffer += pitch;
    }

Now there's only one addition per loop iteration, and no more widening
operations.

So in summary: this clang-tidy check can help eliminate overflow
problems (though maybe rare in practice?). But it can also help generate
(slightly) faster code in some cases.

--

Unfortunately currently this clang-tidy check produces many
false-positive warnings. I opened an issue:
    https://github.com/llvm/llvm-project/issues/58709

--

This patch series fixes many (but not all) of these clang-tidy warnings.
At the same time it also tries to fix some warnings about
narrowing-conversions. See commit 3dd9645450b5 for more details.