Simplify usage of FileOperations::getStat()
Replace the bool-return-value and 'Stat'-output-parameter with a
'std::optional<Stat>' return value (without any output-parameters).
  before: bool getStat(zstring_view filename, Stat& result);
  after:  std::optional<Stat> getStat(zstring_view filename);

This makes it impossible to use an uninitialized 'Stat'. It also
eliminates clang-tidy warnings about _possibly_ uninitialized 'Stat'
structures (because this must be defined (typically uninitialized)
before the old getStat() function can be called). I don't think we had
any such bugs, but it's anyway better to enforce this property in the
API.