Updates build process for macOS
Updates the build process for macOS.

The folder ‘share’ in the app bundle has been relocated to Content/Resources. This allows codesign to work without errors and enhances compatibility for Macs with Apple Silicon.

Furthermore, the app.mk has been adapted to automatically sign the app bundle in case the user has a Apple Developers ID.