bugfix in SaI3xScaler after b0b3d6bcd1 (Dec 2020)
Commit b0b3d6bcd1 changed 'traditional' for-loops into range-based for
loops using xrange(). There was a mistake in SaI3xScaler, it contained
this loop:
    for (unsigned dstY = dstStartY; dstY < dstEndY; srcY++) { ...
It easy to mistake this for a 'traditional' loop. Did you notice what's
"wrong" with it?

Anyway, bug fixed now. And to make this atypical pattern stand out, I
wrote it now using a while-loop (not a for-loop).

This code got broken almost 2 years ago, but nobody noticed (or bothered
to report). I also only noticed by accident now while testing another
refactoring.