Fix XML-entity replacement in machine/extension configs
There are two aspects of our XML parser that combine and triggered a
bug:

First: our 'rapidsax' XML parser can process strings in two modes:
* Pass std::string_view objects via callbacks.
* Same as above, but in addition place a 0-terminator after the string.

Many string algorithms need to know, or can work faster, if the length
of the string is known. So usually it's a good idea to store the full
string_view object. Internally string_view contains a pointer to the
actual data plus a length of this data.

Though in some cases having a more compact representation can be more
important. In that case we can only store a pointer to the start of the
string, and then have the length of the string stored indirectly by
zero-terminating the string. In other words: store it as a C-style
string.

Second: our 'rapidsax' XML parser operates in-place on the XML data.
That is: we read the file in memory, and then change this data in-place.

For example when we perform XML-entity-replacement (stuff like &lt;
or &amp;) we rewrite the strings inside their original location (luckily
these replacements only make the strings shorter, so we always have room
to transform in-place).

A second replacement is to (optionally) zero-terminate the strings. For
example in "<tag>data</tag>", zero-terminating "data" will overwrite the
'<' character of the closing tag. But that's OK, after parsing we don't
need those tags anymore.

Though there was a bug:
When zero-terminating is enabled, and when we do entity-replacement, the
zero-character was wrongly placed in the original location (e.g. the '<'
character of the closing tag), not in a possibly earlier position
because the string has become shorter.

In openMSX we only hit this bug when parsing machine/extension XML
files. Not when parsing the software database or the settings.xml file.
Because only the former use the zero-termination option.

This bug got introduced in commit 1240f5058adfc4 (and follow-ups) which
changes the internal representation of the XMLElement class.