Added "-scrollspeed" property
The intended use case is to make too wide text scroll left and right so that it
becomes fully readable (over time).
Because it's meant for text, only horizontal scrolling is implemented (even
though the property exists for both "text" and "rectangle" widgets).

At a more technical level: the scroll-animation is calculated using "local"
geometry. That is: the rendered width of the current widget (typically text),
the bounding box (width) of the _immediate_ parent widget, and the elapsed time
is taken into account, nothing else.

Patch created by Wouter (on my request), but he asked me to commit it as he was
busy with other things.