Fix crash when video system could not be initialized
OpenMSX crashed instead of showing an error message when the video
system could not be initialized.

(Depening on the system you're using) can artificially be reproduced
with something like:
    SDL_VIDEODRIVER=fb openmsx

The root cause was commit 9087322274904. This marked the 'signalEvent()'
method 'noexcept' (which promises it won't throw exceptions). I did an
audit of the code, and this promise was true, except for the case of
failing to initialize the video system.

I reverted this part of commit 9087322274904 (but left other parts in
tact). See that commit message for more info.