Issue#1461: Fix loading old savestates after e87ef72
Commit e87ef72 changed (corrected) the type of some members from 'int'
to 'unsigned'. Unfortunately that means that old savestates that happen
to contain a negative value for that member now refuse to load.

Fixed by making the parser more lenient. If parsing an 'unsigned' fails,
we now retry parsing as an 'int'.

See: https://github.com/openMSX/openMSX/issues/1461
Thanks a lot to 'mcolom' for reportig this issue!