Show detailed spec previews of machines in OSD menu.
For this we use a temporary created machine. To avoid these temporary machines
triggering all kinds of messages (e.g. about SRAM content, PSG directions,
etc.), I added a setting to suppress these in MSXCliComm. This setting is
really only intended for this purpose (for now).

The script also checks whether the machine can be actually loaded and if not,
shows the error message in the preview info window as well, with a red color to
make it clear the machine won't be running if selected.