Fix -Wconversion warnings in CPU related code
The Z80/R800 emulation code does a lot of bit-width conversions. For
example add 2 8-bit values using 16-bit precision (9-bit would be
sufficient) to be able to calculate the carry flag, then narrow back to
8-bit. This patch makes those narrowing-conversions explicit.

This is still part of evaluating whether enabling -Wconversion by
default is worth it. The CPU emulation code does a lot of type
conversions, so if the syntactical overhead in this code is acceptable,
it will likely also be acceptable in other code.

In many cases this patch makes implicit conversion explicit, which is
good (IMHO). But here is a counter example:

    byte f = 0;
    ...
    f |= table.ZS[getA()];            // no warning
    f |= getF() & (X_FLAG | Y_FLAG);  // does produce a -Wconversion warning

The last two lines OR 8-bit values, the result will always still fit
in 8-bit. In the first case gcc agrees and does not generate a warning.
But the second case is too complex and there gcc does produce a
(false-positive) warning. This patch solves that by rewriting the line
as:
    f |= byte(getF() & (X_FLAG | Y_FLAG));  // no warning anymore

The reason for this is a weird wrinkle(*1) in the C/C++ language. I'm
simplifying a bit: it states that all operations on types smaller than
'int' are first promoted to 'int'. Thus
    f |= getF() & (X_FLAG | Y_FLAG);
is actually evaluated as:
    int tmp1 = int(getF());               // promote 'getf()' to int
    int tmp2 = int(X_FLAG) | int(Y_FLAG); // promote '{X,Y}_FLAG' to int
    int tmp3 = tmp1 & tmp2;
    int tmp4 = int(f) | tmp3;             // promote 'f' to int
    f = byte(tmp4);                       // narrow back to 8-bit
The warning is about that last narrowing conversion (which is harmless,
but gcc isn't smart enough to see this).

Note that the compiler's optimizer will (likely) later eliminate all the
promotion/narrowing conversions again. But this only happens in a later
phase when all front-end warning have already been generated.

I expect that future gcc versions will eliminate this warning, but for
now we need the extra cast.

(*1)
This may seem like a very weird rule, but it does make the following
code work as expected:
    uint8_t x = 100;
    uint8_t y = 200;
    int z = x + y; // result is '300', not '44'

--

Some 16-bit calculations are emulated using 32-bit. This is because on
many host CPU architectures (e.g. x86 and ARM) 32-bit operations are
actually faster than 16-bit. One example are address calculations: Z80
only has a 16-bit address-space, so calculations should wrap at 16-bit.
Still we emulate those with 32-bit (and manually wrap) because later on
those 16-bit Z80 addresses are used as an index in some emulated host
memory buffer, and we save some host instructions (zero-extension from
16-bit) if we immediately start with a wider type. Such extension
instructions are cheap, but we have a lot of such cases, some in
performance critical code.

Compared to ~15 years ago, gcc has gotten better at promoting 16-bit
calculations to 32 or 64-bit calculations (because that's more efficient
on x86/ARM) when it can prove the result will be the same. So
this patch removes _some_ but not all of the manual type promotions
(e.g. port-address are now 16-bit, like on real Z80, but memory
addresses are still emulated using 32-bit, unlike Z80).