Extend type_via_{keyboard,keybuf} with -cancel option
Was an easy extension, requested in this MRC post:
  https://www.msx.org/forum/msx-talk/openmsx/cancelling-or-stopping-the-feature-input-text-and-some-more

Though this change is not 100% backwards compatible:
* Before any single argument type command was interpreted as the to-be
  typed text. Even if that single argument was e.g. "-release". (But this
  rule conflicts with the new command "type -cancel".)
* Now the type command follows the normal Tcl conventions:
  - All arguments that start with a '-' are interpreted as options.
  - Unknown options are flagged as errors.
  - Everything after "--" is interpreted as regular (non-option)
    arguments.

Concretely this means that e.g.
  type -bla
used to work (it typed the string "-bla"), but now it generates an error
about an unknown option. To really type this string you must execute the
command:
  type -- -bla
I'll also adapt catapult in a follow-up commit.