MSXtar: Clamp year and seconds to allowed range.
The date’s narrow would assert for files with a modification date before 1980.
The FAT12 range for year is 0-119 relative to 1980 (up to 2099).

Additionally, tm_sec has a range of 0-60 (0-61 in C++98) to account for leap
seconds, whereas the FAT12 range for seconds/2 is 0-29.

See https://github.com/openMSX/openMSX/issues/1472