MSXtar: Correctly process root directory references
When adding a subdirectory to the root directory, openMSX would assert on the
addition of the `..` entry. This points back to the root directory, however the
root can not be represented with a cluster number, so in FAT12 it needs to be
special-cased with cluster number 0.

In opt builds, and before the addition of the narrow casts, given the root
directory sector `sectorToCluster` returns 0x10000001, which is then clipped to
1, an invalid cluster number. By happenstance it still worked since MSX-DOS2
apparently treats both 0 and 1 as a reference to the root directory.

The fix adds a check whether the directory entry should point to the root
directory, and if so uses the special cluster number. Vice versa, similar checks
are added where it looks up the sector number for directory entries.

Fixes https://github.com/openMSX/openMSX/issues/1472