MSXtar: Stop reading directory when entry is 0x00.
When iterating over directory entries, MSXtar would skip over all entries marked
with 0x00 (and 0xe5), and continue until the end of the sector chain is reached.
However the value 0x00 marks the end of the directory (see [1], [2], [3]) and
any directory entries after this marker are invalid just like deleted ones (0xe5)
or can contain arbitrary garbage, which openMSX should not list or export.

Additionally, in the disk image of issue #1473, the directory entries following
the marker were clean (all zero bytes), however the next cluster in the FAT is
0xFF0 which is an unusual end marker recognised by MS-DOS 3.3 (see [4]).
But since MSX-DOS(2) / Nextor does not treat this cluster number specially
(see [5]), I don’t think MSXtar should either.

Since MSXtar would continue past the 0x00 marker this caused it to access this
cluster 0xFF0, and strange things (a crash) would ensue, while DIR in MSX-DOS(2)
would ignore it because it will have stopped before it reached this point.

This patch exits the loops in `dir` and `recurseDirExtract` when it encounters
the 0x00 marker, as per MSX-DOS and the FAT12 description.

Additionally, it also adds a check for LFN entries in `recurseDirExtract`, which
it should not attempt to extract. This check already existed in `dir`.

Fixes https://github.com/openMSX/openMSX/issues/1473

[1]: https://en.wikipedia.org/wiki/Design_of_the_FAT_file_system#Directory_entry
[2]: https://github.com/Konamiman/Nextor/blob/v2.1/source/kernel/bank2/dir.mac#L308
[3]: https://github.com/Konamiman/Nextor/blob/v2.1/source/kernel/bank3/dos1ker.mac#L842
[4]: https://en.wikipedia.org/wiki/Design_of_the_FAT_file_system#Cluster_values
[5]: https://github.com/Konamiman/Nextor/blob/v2.1/source/kernel/bank2/fat.mac#L62