MSXtar: Assert for valid cluster numbers
Add checks to `clusterToSector` and `sectorToCluster` to ensure the cluster
number it consumes or produces is a valid one that can be converted to a sector
and vice versa. Since they do not produce error values the inputs for these are
supposed to be checked prior to the call to guarantee a valid result.

Also cap the max cluster to BAD_FAT, as MSX-DOS(2) / Nextor can only handle
cluster numbers below that (see [1]).

See https://github.com/openMSX/openMSX/issues/1472

[1]: https://github.com/Konamiman/Nextor/blob/v2.1/source/kernel/bank2/fat.mac#L62