MSXtar: Do not crash on FREE_FAT in FAT chain.
When scanning a FAT chain, when it encounters FREE_FAT this is a file system
integrity error. After all, this cluster could be allocated later for a
different chain, and they would then be conjoined.

Either way, this file system problem is not a reason to assert, so we check.

This issue is encountered when exporting the root of the disk image attached in:
https://github.com/openMSX/openMSX/issues/1473