MSXtar: Store clusterCount instead of maxCluster.
Having `maxCluster` which included the `FIRST_CLUSTER` offset (+ 2) made it
confusing to reason about the ranges that it applied to. Additionally, it was
unclear whether it was inclusive or exclusive. The new `clusterCount` purely
talks about the amount of clusters, and since it’s referring to a count, it’s
clear that it is exclusive.