MSXtar: Don’t check cluster count in clusterToSector.
This check caused a subtle edge case crash on debug builds in getNextSector(),
which compares sectorToCluster(sector) to sectorToCluster(sector + 1). When this
exceeds the cluster count the assertion will trigger unnecessarily.

Since values are normalised when read, and there are checks when writing cluster
numbers, there should be plenty of protection elsewhere.