Fix running binary in app bundle through a symlink.
In addition to searching for a `share` directory, also search for
`Contents/Resources/share`.

This fixes a regression from commit 739918dd22, where running openMSX from the
command line through a symlink to `openMSX.app/Contents/MacOS/openmsx` would not
be able to find the `Contents/Resources/share` directory in the app bundle
(previously it was in the root of the bundle so it would be found).

Unfortunately `[NSBundle mainBundle]` does not expand symlinks and therefore
does not recognise a proper application bundle (it does without symlink), so
`share` isn’t found by the `[bundle URLForResource]` call done earlier.

Test case, with openMSX.app installed in `/Applications`:

    ln -s /Applications/openMSX.app/Contents/MacOS/openmsx /usr/local/bin/openmsx
    openmsx

See https://github.com/openMSX/openMSX/pull/1457